requirements.txt
colorama
cx-Oracle
Flask
Flask-RESTful
Flask-SQLAlchemy
pytest
# Web Service Client
suds-jurko